% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter_ra.R
\name{converter_ra}
\alias{converter_ra}
\title{Check brazilian documents.}
\usage{
converter_ra(entrada)
}
\arguments{
\item{entrada}{Caracter or numeric of the document that will be validated.

#' @return True or False. if log is giving return a data frame.}
}
\description{
\code{converter_ra} returns a brazilian document without special characters and as a numeric.
}
\examples{
 converter_ra("529.982.247-25")
}

