% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter_ra.R
\name{doc2integer64}
\alias{doc2integer64}
\title{Check brazilian documents.}
\usage{
doc2integer64(entrada)
}
\arguments{
\item{entrada}{Caracter or numeric of the document that will be validated.}
}
\value{
a bit64 integer
}
\description{
\code{doc2integer64} returns a brazilian document without special characters and as a numeric.
}

