% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrai_NomeProprio.R
\name{extrai_NomeProprio}
\alias{extrai_NomeProprio}
\title{Check Brazilian names}
\usage{
extrai_NomeProprio(x, sobrenome = FALSE, sexo = FALSE)
}
\arguments{
\item{x}{List, character or factor with names to be parsed.}

\item{sobrenome}{If TRUE, the list of surnames is returned.}

\item{sexo}{If TRUE, the list of gender based on the names is returned.

importFrom data.table stringr
return Returns a data.table}
}
\description{
\code{extrai_NomeProprio} Parse brazlian names and returns given names, surnames and gender
}
\examples{
extrai_NomeProprio(x = c("Maria Conceicao da Costa", "Mario Silva"), sobrenome = TRUE, sexo = FALSE)
}

