% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valida_documento.R
\name{graficoDOC}
\alias{graficoDOC}
\title{Check brazilian documents.}
\usage{
graficoDOC(base = NULL, data = NULL, columns, types, filename)
}
\arguments{
\item{data}{Data frame with data to be validate.}

\item{columns}{Column of data that will be evaluated.}

\item{types}{Caracter, it could be cpf, cnpf, pis e titulo de eleitor.}
}
\value{
True or False. if log is giving return a data frame.
}
\description{
\code{graficoDOC} returns true or false if the number of document is correct.
}

