% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valida_documento.R
\name{relatorioDOC}
\alias{relatorioDOC}
\title{Check brazilian documents.}
\usage{
relatorioDOC(input_file = NULL, data = NULL, columns, types,
  output_filename, tipo_relatorio = "tabela")
}
\arguments{
\item{input_file}{file where the data is located.}

\item{data}{Data frame with data to be validate.}

\item{columns}{Columns of data that will be evaluated.}

\item{types}{Caracter, it could be cpf, cnpf, pis e titulo de eleitor.}

\item{output_filename}{character, where the output will be save.}

\item{tipo_relatorio}{character, grafico ou tabela, indicates the kind of output}
}
\value{
return a data frame.
}
\description{
\code{relatorioDOC} output on pdf a data frame that summarize the output of validaDOC.
                    It is require a file where the data is located or the data itself.
                    If neither, input and data, is giving the function stop. Also,
                    columns and types must the same size.
}
