% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tratadocumento.R
\name{trata_cpf}
\alias{trata_cpf}
\title{Correct Brazilian cpf document.}
\usage{
trata_cpf(entrada)
}
\arguments{
\item{entrada}{Character or numeric of the document that will be adjusted.}
}
\value{
Entrada adjusted.
}
\description{
\code{trada_cpf} The aim is transform the entrada to the pattern of Brazilian document cpf.
                 It must have 11 numeric characters. When a dataset is read as a numeric,
                 it lost the 0 character on the left. This function fix that lost and transform
                 entrada to character.
}
\examples{
   trata_cpf(52145795)
}
