% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter_ra.R
\name{doc2integer64}
\alias{doc2integer64}
\title{Convert brazilian documents.}
\usage{
doc2integer64(entrada)
}
\arguments{
\item{entrada}{Caracter or numeric that will be converted.}
}
\value{
an bit64 integer
}
\description{
\code{doc2integer64} returns a brazilian document without special characters and as an integer64.
}

