% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera_digitos.R
\name{gera_digito}
\alias{gera_digito}
\title{Check brazilian documents.}
\usage{
gera_digito(entrada, type = "cpf")
}
\arguments{
\item{entrada}{Caracter or numeric of the document that will be validated.}

\item{type}{Caracter, it could be cpf, cnpf, pis e titulo de eleitor.}
}
\value{
it will alter the number by reference.
}
\description{
\code{gera_digito} generate the last digit of a id.
}
\examples{
 gera_digito("529.982.247-2", type = "cpf")
 gera_digito("60.149.443/0001-7", type = "cnpj")
}
