% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tratadocumento.R
\name{trata_cpf}
\alias{trata_cpf}
\title{Check brazilian documents.}
\usage{
trata_cpf(entrada)
}
\arguments{
\item{entrada}{Caracter or numeric of the document that will be validated.}

\item{type}{Caracter, it could be cpf, cnpf, pis e titulo de eleitor.}

\item{log}{Output erros found on entrada}
}
\value{
True or False. if log is giving return a data frame.
}
\description{
\code{trada_cpf} returns true or false if the number of document is correct.
}

