% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valida_documento.R
\name{valida_doc}
\alias{valida_doc}
\title{Check brazilian documents.}
\usage{
valida_doc(entrada, type = "cpf")
}
\arguments{
\item{entrada}{Caracter or numeric of the document that will be validated.}

\item{type}{Caracter, it could be cpf, cnpf, pis e titulo de eleitor.}
}
\value{
True or False.
}
\description{
\code{valida_doc} returns true or false if the number of document is correct.
}
\examples{
 valida_doc("529.982.247-25", type = "cpf") 
 valida_doc("60.149.443/0001-70", type = "cnpj")
}

