% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valida_documento.R
\name{valida_doc_df}
\alias{valida_doc_df}
\title{Check Brazilian documents.}
\usage{
valida_doc_df(data, column, type = "cpf", log = FALSE)
}
\arguments{
\item{data}{Data frame with data to be validate.}

\item{column}{Column of data that will be evaluated.}

\item{type}{Caracter, it could be cpf, cnpf, pis e titulo de eleitor.}

\item{log}{Output erros found on entrada}
}
\value{
True or False. if log is giving return a data frame.
}
\description{
\code{valida_doc_df} returns true or false if the number of document is correct.
}
